requirements.txt
azure_storage
beautifulsoup4
pandas
requests
